+++ src/main/java/com/heartfoilo/demo/domain/invest/entity/Order.java
@@ -35,7 +35,7 @@ public class Order {
 
 
     @Column(name = "order_amount", nullable = false)
-    private int orderAmount;
+    private Long orderAmount;
 
     @Column(name = "order_price", nullable = false)
     private int orderPrice;
